Wait 2 seconds for `daemonize` write the PID file
`daemonize` forks and immediately exits. The forked process writes its
PID into the file, then `exec()`utes the target program.

The script that runs `daemonize` waits for it to exit, checks the exit
code, and reads the file. But on a heavily loaded system, `daemonize`
might exit *before* its child wrote its PID to a file. As a result, our
main script fails to read the PID and crashes.