Implement .plentyfs/seed
Fixes #13.

I put the file into a dot-directory because I'm not sure if it will be
the only such file. Perhaps all the mount options should be reflected
there? I don't know yet.

Subplot that checked the contents of the file had to be adjusted: this
commit changed the inodes of the files and thus their contents.

I haven't yet formed a vision of how to best structure this code, so
I continue to pile it up.